Implement multi-role online mall

+++ src/main/java/com/shop/strategy/DiscountStrategy.java
@@ -1,4 +1,7 @@
 package com.shop.strategy;
+
+import java.math.BigDecimal;
+
 public interface DiscountStrategy {
-    double calculate(double originalPrice, double discountValue);
+    BigDecimal calculate(BigDecimal originalPrice, BigDecimal discountValue);
 }
